+++ packages/moonstone/VideoPlayer/VideoPlayer.js
@@ -609,7 +609,12 @@ const VideoPlayerBase = class extends React.Component {
 
 		if (!compareSources(source, nextSource)) {
 			this.firstPlayReadFlag = true;
-			this.setState({currentTime: 0, proportionPlayed: 0, proportionLoaded: 0});
+			this.setState({
+				announce: AnnounceState.READY,
+				currentTime: 0,
+				proportionPlayed: 0,
+				proportionLoaded: 0
+			});
 		}
 	}
 
@@ -663,6 +668,7 @@ const VideoPlayerBase = class extends React.Component {
 		if (!compareSources(source, prevSource)) {
 			const isPreloadedVideo = source && prevPreloadSource && compareSources(source, prevPreloadSource);
 			this.reloadVideo(isPreloadedVideo);
+			this.showControls();
 		}
 
 		this.setFloatingLayerShowing(this.state.mediaControlsVisible || this.state.mediaSliderVisible);
@@ -891,9 +897,12 @@ const VideoPlayerBase = class extends React.Component {
 	}
 
 	showFeedback = () => {
-		if (this.state.mediaControlsVisible && !this.state.feedbackVisible) {
-			this.setState({feedbackVisible: true});
-		} else if (!this.state.mediaControlsVisible) {
+		if (this.state.mediaControlsVisible) {
+			this.setState({
+				feedbackIconVisible: true,
+				feedbackVisible: true
+			});
+		} else {
 			const shouldShowSlider = this.pulsedPlaybackState !== null || calcNumberValueOfPlaybackRate(this.playbackRate) !== 1;
 
 			if (this.showMiniFeedback && (!this.state.miniFeedbackVisible || this.state.mediaSliderVisible !== shouldShowSlider)) {
@@ -1051,6 +1060,7 @@ const VideoPlayerBase = class extends React.Component {
 	}
 
 	renderBottomControl = new Job(() => {
+		this.showControls();
 		this.setState({bottomControlsRendered: true});
 	});
 
@@ -1469,7 +1479,7 @@ const VideoPlayerBase = class extends React.Component {
 			this.player.querySelector(
 				`.${css.leftComponents} ${defaultSpottable}, .${css.rightComponents} ${defaultSpottable}`
 			) ||
-			'data-media-controls';
+			this.player.querySelector(`[data-media-controls] ${defaultSpottable}`);
 
 		return defaultControl ? Spotlight.focus(defaultControl) : false;
 	}
@@ -1535,7 +1545,8 @@ const VideoPlayerBase = class extends React.Component {
 		this.sliderScrubbing = false;
 		this.startDelayedFeedbackHide();
 		this.setState({
-			feedbackIconVisible: true,
+			// If paused is false that means it is playing. We only want to hide on playing.
+			feedbackIconVisible: this.state.paused,
 			sliderTooltipTime: this.state.currentTime
 		});
 	}