+++ packages/spotlight/Spottable/Spottable.js
@@ -10,6 +10,7 @@ import {forward, forwardWithPrevent, handle} from '@enact/core/handle';
 import hoc from '@enact/core/hoc';
 import {is} from '@enact/core/keymap';
 import React from 'react';
+import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
 
 import {getContainersForNode} from '../src/container';
@@ -43,6 +44,8 @@ const isKeyboardAccessible = (node) => {
 		)
 	);
 };
+
+const isSpottable = (props) => !props.disabled && !props.spotlightDisabled;
 
 // Last instance of spottable to be focused
 let lastSelectTarget = null;
@@ -188,21 +191,26 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 			this.isHovered = false;
 
 			this.state = {
-				spottableDisabled: false
+				focusedWhenDisabled: false
 			};
 		}
 
-		componentWillReceiveProps (nextProps) {
-			const spottableDisabled = this.isFocused && nextProps.disabled || nextProps.spotlightDisabled;
+		componentDidMount () {
+			// eslint-disable-next-line react/no-find-dom-node
+			this.node = ReactDOM.findDOMNode(this);
+		}
 
-			if (!this.state.spottableDisabled && spottableDisabled) {
-				this.setState({spottableDisabled});
-			}
+		componentWillReceiveProps (nextProps) {
+			const focusedWhenDisabled = this.isFocused && nextProps.disabled && !this.props.disabled;
+
+			this.setState({
+				focusedWhenDisabled
+			});
 		}
 
 		componentDidUpdate (prevProps, prevState) {
 			// if the component is focused and became disabled
-			if (!prevState.spottableDisabled && this.state.spottableDisabled) {
+			if (!prevState.focusedWhenDisabled && this.state.focusedWhenDisabled) {
 				if (lastSelectTarget === this) {
 					selectCancelled = true;
 					forward('onMouseUp', null, this.props);
@@ -210,12 +218,7 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 			}
 
 			// if the component became enabled, notify spotlight to enable restoring "lost" focus
-			if (
-				(!this.props.disabled && !this.props.spotlightDisabled) && (
-					(prevProps.disabled && !this.props.disabled) ||
-					(prevProps.spotlightDisabled && !this.props.spotlightDisabled)
-				)
-			) {
+			if (isSpottable(this.props) && !isSpottable(prevProps)) {
 				if (Spotlight.getPointerMode()) {
 					if (this.isHovered) {
 						Spotlight.setPointerMode(false);
@@ -289,9 +292,7 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 			return notPrevented && allow;
 		}
 
-		isActionable = (ev, props) => {
-			return !props.disabled && !props.spotlightDisabled;
-		}
+		isActionable = (ev, props) => isSpottable(props)
 
 		handle = handle.bind(this)
 
@@ -316,8 +317,8 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 			if (ev.currentTarget === ev.target) {
 				this.isFocused = false;
 
-				if (this.state.spottableDisabled) {
-					this.setState({spottableDisabled: false});
+				if (this.state.focusedWhenDisabled) {
+					this.setState({focusedWhenDisabled: false});
 				}
 			}
 
@@ -363,8 +364,8 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		render () {
-			const {disabled, spotlightDisabled, spotlightId, ...rest} = this.props;
-			const spottable = this.state.spottableDisabled || !disabled && !spotlightDisabled;
+			const {disabled, spotlightId, ...rest} = this.props;
+			const spottable = this.state.focusedWhenDisabled || isSpottable(this.props);
 			let tabIndex = rest.tabIndex;
 
 			delete rest.onSpotlightDisappear;
@@ -372,6 +373,7 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 			delete rest.onSpotlightLeft;
 			delete rest.onSpotlightRight;
 			delete rest.onSpotlightUp;
+			delete rest.spotlightDisabled;
 
 			if (tabIndex == null) {
 				tabIndex = -1;